Complete C++ plumbing for scattering materials
Former-commit-id: f1ff71e280d8f166f38dfd2749393daf64e0a7fd

+++ ox_wrapper/ox_wrapper/miss_shader.cpp
@@ -28,5 +28,5 @@ bool OxMissShader::isValid() const
 
 void OxMissShader::apply() const
 {
-    throwOptiXContextError(rtContextSetMissProgram(nativeOptiXContextHandle(), m_ray_type, nativeOptiXProgramHandle()));
+    throwOptiXContextError(rtContextSetMissProgram(nativeOptiXContextHandle(), static_cast<unsigned int>(m_ray_type), nativeOptiXProgramHandle()));
 }
